requirements.txt
Django==5.2.4
djangorestframework==3.16.0
psycopg2==3.2.9
PyJWT==2.10.1
django-cors-headers==4.7.0